+++ src/main/java/auth/support/AdminInterceptor.java
@@ -1,5 +1,6 @@
 package auth.support;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.servlet.HandlerInterceptor;
 
@@ -7,12 +8,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.util.Objects;
 
+@RequiredArgsConstructor
 public class AdminInterceptor implements HandlerInterceptor {
-    private JwtTokenProvider jwtTokenProvider;
-
-    public AdminInterceptor(JwtTokenProvider jwtTokenProvider) {
-        this.jwtTokenProvider = jwtTokenProvider;
-    }
+    private final JwtTokenProvider jwtTokenProvider;
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
