+++ src/main/java/nextstep/controller/ReservationController.java
@@ -3,10 +3,9 @@ package nextstep.controller;
 import auth.domain.persist.UserDetails;
 import auth.support.AuthenticationException;
 import auth.domain.template.LoginMember;
-import nextstep.domain.dto.ReservationResponse;
-import nextstep.domain.persist.Member;
+import nextstep.domain.dto.response.ReservationResponse;
 import nextstep.domain.persist.Reservation;
-import nextstep.domain.dto.ReservationRequest;
+import nextstep.domain.dto.request.ReservationRequest;
 import nextstep.service.ReservationService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
