+++ src/main/java/nextstep/domain/dto/request/ScheduleRequest.java
@@ -1,4 +1,4 @@
-package nextstep.domain.dto;
+package nextstep.domain.dto.request;
 
 import nextstep.domain.persist.Schedule;
 import nextstep.domain.persist.Theme;
