+++ src/main/java/nextstep/domain/dto/request/WaitingRequest.java
@@ -1,4 +1,4 @@
-package nextstep.domain.dto;
+package nextstep.domain.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
