+++ src/main/java/nextstep/service/WaitingService.java
@@ -3,8 +3,8 @@ package nextstep.service;
 import auth.domain.persist.UserDetails;
 import auth.support.AuthenticationException;
 import lombok.RequiredArgsConstructor;
-import nextstep.domain.dto.WaitingRequest;
-import nextstep.domain.dto.WaitingResponse;
+import nextstep.domain.dto.request.WaitingRequest;
+import nextstep.domain.dto.response.WaitingResponse;
 import nextstep.domain.persist.Member;
 import nextstep.domain.persist.Reservation;
 import nextstep.domain.persist.Waiting;
@@ -29,21 +29,26 @@ public class WaitingService {
     public String create(UserDetails userDetails, WaitingRequest waitingRequest) {
         Long scheduleId = waitingRequest.getScheduleId();
         List<Reservation> reservations = reservationDao.findByScheduleId(scheduleId);
+
         if (reservations.isEmpty()) {
-            Long id = reservationDao.save(new Reservation(scheduleDao.findById(scheduleId), new Member(userDetails)));
-            return "/" + RESERVATION + "/" + id;
+            Long reservationId = reservationDao.save(new Reservation(scheduleDao.findById(scheduleId), new Member(userDetails)));
+            return "/" + RESERVATION + "/" + reservationId;
         }
-        Long id = waitingDao.save(new Waiting(scheduleDao.findById(scheduleId), new Member(userDetails)));
-        return "/" + WAITING + "/" + id;
+
+        Long waitingId = waitingDao.save(new Waiting(scheduleDao.findById(scheduleId), new Member(userDetails)));
+        return "/" + WAITING + "/" + waitingId;
     }
 
     public List<WaitingResponse> findAll(Long id) {
-        // 해당 memberId가 예약한 scheduleId, waitingId 가져옴
         List<Waiting> waitings = waitingDao.findAll(id);
-        // scheduleId에 해당하는 waitingId의 순번을 가져옴
+
         return waitings.stream()
-                .map(r -> (new WaitingResponse(r, waitingDao.getPriority(r.getSchedule().getId(), r.getId()))))
+                .map(waiting -> (new WaitingResponse(waiting, getPriority(waiting))))
                 .collect(Collectors.toList());
+    }
+
+    private Long getPriority(Waiting waiting) {
+        return waitingDao.getPriority(waiting.getSchedule().getId(), waiting.getId());
     }
 
     public void deleteById(UserDetails userDetails, Long id) {
