+++ src/test/java/nextstep/auth/AuthE2ETest.java
@@ -3,8 +3,8 @@ package nextstep.auth;
 import auth.domain.dto.TokenRequest;
 import auth.domain.dto.TokenResponse;
 import io.restassured.RestAssured;
-import nextstep.domain.dto.MemberRequest;
-import nextstep.domain.dto.ThemeRequest;
+import nextstep.domain.dto.request.MemberRequest;
+import nextstep.domain.dto.request.ThemeRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
