+++ src/test/java/nextstep/reservation/ReservationE2ETest.java
@@ -4,10 +4,10 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.AbstractE2ETest;
-import nextstep.domain.dto.ReservationRequest;
-import nextstep.domain.dto.ReservationResponse;
-import nextstep.domain.dto.ScheduleRequest;
-import nextstep.domain.dto.ThemeRequest;
+import nextstep.domain.dto.request.ReservationRequest;
+import nextstep.domain.dto.response.ReservationResponse;
+import nextstep.domain.dto.request.ScheduleRequest;
+import nextstep.domain.dto.request.ThemeRequest;
 import nextstep.domain.persist.Reservation;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
