+++ src/test/java/nextstep/theme/ThemeE2ETest.java
@@ -4,8 +4,8 @@ import io.restassured.RestAssured;
 import nextstep.AbstractE2ETest;
 import auth.domain.dto.TokenRequest;
 import auth.domain.dto.TokenResponse;
-import nextstep.domain.dto.MemberRequest;
-import nextstep.domain.dto.ThemeRequest;
+import nextstep.domain.dto.request.MemberRequest;
+import nextstep.domain.dto.request.ThemeRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
