+++ src/test/java/nextstep/waiting/WaitingE2ETest.java
@@ -4,7 +4,11 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.AbstractE2ETest;
-import nextstep.domain.dto.*;
+import nextstep.domain.dto.request.ReservationRequest;
+import nextstep.domain.dto.request.ScheduleRequest;
+import nextstep.domain.dto.request.ThemeRequest;
+import nextstep.domain.dto.response.ReservationResponse;
+import nextstep.domain.dto.response.WaitingResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
